+++ resources/views/shops/skg-mall.blade.php
@@ -77,14 +77,3 @@ SKG MALL
 {{-- page level scripts --}}
 @section('footer_scripts')
 @stop
-
-
-
-
-
-
-
-
-
-0
-0
